+++ inst/examples/ensureExampleFiles.r
@@ -1,5 +1,6 @@
 \dontrun{
 
 ensureExampleFiles()
+ensureExampleFiles(strict = TRUE)
 
 }